vylepšení ortogonalizace - poslední prvek
parametry - drobná úprava výpisu
my - výpočet azimutu misto float double

+++ my.cpp
@@ -93,7 +93,7 @@ double Cmy::azimut(double X1,double Y1,double X2,double Y2)
 		try{
 			if(delka(X1,Y1,X2,Y2)!=0)
 			{
-				float A=acos((Y2-Y1)/(delka(X1,Y1,X2,Y2)*1.0))*180/M_PI;
+				double A=acos((Y2-Y1)/(delka(X1,Y1,X2,Y2)*1.0))*180/M_PI;
 				if(X2<X1)A=360-A;
 				return A;
 			}
